First upload with Template files
Contains an Index, Template, four includes (three of which are currently implemented, one of which will be only on the form pages), and three CSS stylesheets. As of this upload, I do not have LESS implemented into my website yet. However, this should be the only missing feature. I need to brush up on my LESS in the next few days, and it should be implemented by the end of the week.